backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   starlette
    #   watchfiles
boto3==1.37.6
    # via pymongo-auth-aws
botocore==1.37.6
    # via
    #   boto3
    #   pymongo-auth-aws
    #   s3transfer
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via pydantic
fastapi==0.115.11
    # via -r requirements.in
filelock==3.17.0
    # via torch
fsspec==2025.2.0
    # via torch
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   requests
jinja2==3.1.5
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.4.2
    # via scikit-learn
markupsafe==3.0.2
    # via jinja2
motor[aws]==3.7.0
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.3
    # via
    #   scikit-learn
    #   scipy
    #   skorch
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
pydantic[email]==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
pymongo[aws]==4.11.2
    # via motor
pymongo-auth-aws==1.3.0
    # via pymongo
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via uvicorn
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==2.0.0
    # via tweepy
s3transfer==0.11.4
    # via boto3
scikit-learn==1.6.1
    # via
    #   -r requirements.in
    #   skorch
scipy==1.15.2
    # via
    #   scikit-learn
    #   skorch
six==1.17.0
    # via python-dateutil
skorch==1.1.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.46.0
    # via fastapi
sympy==1.13.1
    # via torch
tabulate==0.9.0
    # via skorch
threadpoolctl==3.5.0
    # via scikit-learn
torch==2.6.0
    # via -r requirements.in
tqdm==4.67.1
    # via skorch
tweepy==4.15.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via
    #   botocore
    #   requests
uvicorn[standard]==0.34.0
    # via -r requirements.in
watchfiles==1.0.4
    # via uvicorn
websockets==15.0
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
